Backend/requirements.txt
# Django + DRF
Django==5.2.6
djangorestframework==3.16.1
django-cors-headers==4.4.0

# AWS S3
boto3==1.40.37
botocore==1.40.37

# Qdrant Vector DB
qdrant-client==1.15.1

# PDF / DOCX parsing
PyMuPDF==1.26.4
python-docx==1.2.0

# Embeddings
sentence-transformers==5.1.1
torch>=1.13.0
transformers>=4.35.0

# Gemini (Google Generative AI)
google-generativeai==0.7.1

# Environment vars
python-dotenv==1.0.1

# Utilities
requests==2.32.3
numpy>=1.26.0
